refactor: refactor form to refs
