+++ src/pages/Pages.test.tsx
@@ -1,0 +1,29 @@
+import React from 'react';
+import { render, screen, cleanup } from '@testing-library/react';
+import { About } from 'pages/About';
+import { NotFoundPage } from 'pages/NotFound';
+import { BrowserRouter } from 'react-router-dom';
+
+afterEach(cleanup);
+
+describe('Pages tests', function () {
+  test('should render about page', () => {
+    render(<About />);
+
+    const header = screen.getByRole('heading');
+    expect(header).toBeInTheDocument();
+    expect(header).toHaveTextContent('Some info about us!');
+  });
+
+  test('should render 404 page', () => {
+    render(
+      <BrowserRouter>
+        <NotFoundPage />
+      </BrowserRouter>
+    );
+
+    const header = screen.getByRole('heading');
+    expect(header).toBeInTheDocument();
+    expect(header).toHaveTextContent('Oops!');
+  });
+});
